+++ sources/FtpsServerAvalonia/FtpsServerAvalonia.Android/MainActivity.cs
@@ -5,6 +5,7 @@ using Android.OS;
 using Android.Runtime;
 using Android.Views;
 using Avalonia.Android;
+using FtpsServerAvalonia.Services;
 
 namespace FtpsServerAvalonia.Android;
 
@@ -15,10 +16,12 @@ namespace FtpsServerAvalonia.Android;
     MainLauncher = true,
     ConfigurationChanges = ConfigChanges.Orientation | ConfigChanges.ScreenSize | ConfigChanges.UiMode,
     WindowSoftInputMode = SoftInput.AdjustResize)]
-public class MainActivity : AvaloniaMainActivity
+public class MainActivity : AvaloniaMainActivity, IAndroidKeepAwakeService
 {
     protected override void OnCreate(Bundle? savedInstanceState)
     {
+        global::FtpsServerAvalonia.App.AndroidKeepAwakeService = this;
+
         AppDomain.CurrentDomain.UnhandledException += (_, e) =>
         {
             if (e.ExceptionObject is Exception ex)
@@ -38,6 +41,25 @@ public class MainActivity : AvaloniaMainActivity
         {
             ShowErrorDialog(ex);
         }
+    }
+
+    protected override void OnDestroy()
+    {
+        if (global::FtpsServerAvalonia.App.AndroidKeepAwakeService == this)
+            global::FtpsServerAvalonia.App.AndroidKeepAwakeService = null;
+
+        base.OnDestroy();
+    }
+
+    public void SetKeepScreenOn(bool enabled)
+    {
+        RunOnUiThread(() =>
+        {
+            if (enabled)
+                Window?.AddFlags(WindowManagerFlags.KeepScreenOn);
+            else
+                Window?.ClearFlags(WindowManagerFlags.KeepScreenOn);
+        });
     }
 
     private void ShowErrorDialog(Exception ex)